Remove debug card reader module
